requirements.txt
# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# AWS
boto3==1.28.62
awscli==1.29.62
aws-lambda-powertools==2.30.0

# GCP
google-cloud-aiplatform==1.36.4
google-cloud-storage==2.13.0

# Snowflake
snowflake-connector-python==3.4.0
snowflake-sqlalchemy==1.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
tensorflow-recommenders==0.7.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
tqdm==4.66.1
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0 